+++ impli.cabal
@@ -51,27 +51,17 @@ library
   -- cabal-gild: discover src
   exposed-modules:
     Config
-    IMP.Config
+    IMP.Exception
+    IMP.Expression
+    IMP.Lexer
     IMP.Parser
     IMP.Pretty
     IMP.REPL
-    IMP.Result
-    IMP.Semantic.Expression
-    IMP.Semantic.State
-    IMP.Semantic.Statement
+    IMP.Semantic.Operational
+    IMP.Semantic.Structural
+    IMP.State
+    IMP.Statement
     IMP.Syntax
-    IMP.Util
-    IMP2.Exception
-    IMP2.Expression
-    IMP2.Lexer
-    IMP2.Parser
-    IMP2.Pretty
-    IMP2.REPL
-    IMP2.Semantic.Operational
-    IMP2.Semantic.Structural
-    IMP2.State
-    IMP2.Statement
-    IMP2.Syntax
 
 executable impli
   import: shared-settings
