+++ src/IMP/Expression.hs
@@ -9,36 +9,29 @@ Maintainer  : bfeitknecht@ethz.ch
 Stability   : stable
 Portability : portable
 
-This module defines the evaluation semantics for arithmetic and boolean expressions.
-It provides the 'Evaluate' typeclass, which is used by "IMP.Semantic.Statement" to
-evaluate expressions within a program state. The module implements evaluation rules
-for all expression types defined in "IMP.Syntax".
+TODO
 -}
-module IMP.Semantic.Expression (
+module IMP.Expression (
     Evaluate,
     evaluate,
 ) where
 
-import IMP.Semantic.State
+import IMP.State
 import IMP.Syntax
-import IMP.Util
 
--- | Typeclass for evaluating expressions or statements in a state.
+-- | TODO
 class Evaluate a b | a -> b where
-    -- | Evaluate a value of type @a@ in the given state, producing a result of type @b@.
-    --        For 'Aexp' it produces an 'Integer', for 'Bexp' it produces a 'GHC.Types.Bool',
-    --        and for 'Stm' it produces an 'Integer' counting variable definitions.
     evaluate :: State -> a -> b
 
--- | Evaluate an arithmetic expression to an integer.
+-- | TODO
 instance Evaluate Aexp Integer where
     evaluate state aexp = case aexp of
         Val n -> n
         Var x -> getVar state x
-        Bin aop e1 e2 ->
+        Bin aop a1 a2 ->
             let
-                v1 = evaluate state e1
-                v2 = evaluate state e2
+                v1 = evaluate state a1
+                v2 = evaluate state a2
             in
                 case aop of
                     Add -> v1 + v2
@@ -48,17 +41,17 @@ instance Evaluate Aexp Integer where
                     Mod -> v1 %% v2
         Time s -> evaluate state s
 
--- | Evaluate a boolean expression to a boolean value.
+-- | TODO
 instance Evaluate Bexp Bool where
     evaluate state bexp = case bexp of
         Lit b -> b
         Or b1 b2 -> evaluate state b1 || evaluate state b2
         And b1 b2 -> evaluate state b1 && evaluate state b2
         Not b -> not (evaluate state b)
-        Rel rop e1 e2 ->
+        Rel rop a1 a2 ->
             let
-                v1 = evaluate state e1
-                v2 = evaluate state e2
+                v1 = evaluate state a1
+                v2 = evaluate state a2
             in
                 case rop of
                     Eq -> v1 == v2
@@ -68,13 +61,15 @@ instance Evaluate Bexp Bool where
                     Gt -> v1 > v2
                     Geq -> v1 >= v2
 
--- | Evaluate a statement to an integer (number of variable definitions).
+-- | TODO
+-- CHECK
+-- rewrite to count steps in operational semantics
 instance Evaluate Stm Integer where
     evaluate state stm = case stm of
         Skip -> 0
         VarDef {} -> 1
         Seq s1 s2 -> evaluate state s1 + evaluate state s2
-        If b s1 s2 ->
+        IfElse b s1 s2 ->
             if evaluate state b
                 then evaluate state s1
                 else evaluate state s2
@@ -95,19 +90,19 @@ instance Evaluate Stm Integer where
         Return _ _ -> error $ "illegal statement for evaluation: " ++ show stm
         Break -> 0
         Revert s _ -> evaluate state s
-        Match e ms d -> do
-            let v = evaluate state e
+        Match a ms d -> do
+            let v = evaluate state a
             case lookup v ms of
                 Just s -> evaluate state s
                 Nothing -> evaluate state d
         Havoc _ -> 1
         Assert _ -> 0
-        Flip i s1 s2 ->
+        FlipFlop i s1 s2 ->
             if getFlip state i
                 then evaluate state s1
                 else evaluate state s2
         Raise _ -> 0
-        Try s1 _ s2 -> max (evaluate state s1) (evaluate state s2 + 1)
+        TryCatch s1 _ s2 -> max (evaluate state s1) (evaluate state s2 + 1)
         Swap _ _ -> 2
-        Timeout s e -> min (evaluate state s) (evaluate state e)
+        Timeout s a -> min (evaluate state s) (evaluate state a)
         Alternate s1 s2 -> evaluate state s1 + evaluate state s2
