+++ src/IMP/Parser.hs
@@ -9,34 +9,34 @@ Maintainer  : bfeitknecht@ethz.ch
 Stability   : stable
 Portability : portable
 
-This module provides parsers for the constructs defined in "IMP.Syntax".
-For this it utilizes the "Text.Parsec" library and exposes the 'Parse' typeclass and 'parse' function.
-Additionally, the 'parser' function serves as top level abstraction used in "IMP.REPL" and "IMP.CLI".
+TODO
 -}
 module IMP.Parser (
-    Parse,
-    parse,
+    Parses,
+    parses,
     parser,
 ) where
 
-import Text.Parsec hiding (parse)
+import Text.Parsec
 import Text.Parsec.Expr
-import Text.Parsec.String (Parser)
+import Text.Parsec.String
 
-import qualified Text.Parsec.Prim as Parsec
-
+import Config
+import IMP.Lexer
 import IMP.Syntax
-import IMP.Util
 
--- | Typeclass for "IMP.Syntax" data types that can be parsed from source code string
-class Parse a where
-    -- | The parser for the type.
-    parse :: Parser a
+-- | TODO
+parser :: (Parses a) => String -> String -> Either ParseError a
+parser = parse (whitespace *> parses <* eof)
 
--- | Parser for arithmetic expressions, 'Aexp'.
--- Supports numerals, variables, arithmetic operations, and the @time@ operation.
-instance Parse Aexp where
-    parse = buildExpressionParser table term <?> "arithmetic expression"
+-- | TODO
+class Parses a where
+    -- | TODO
+    parses :: Parser a
+
+-- | TODO
+instance Parses Aexp where
+    parses = buildExpressionParser table term <?> "arithmetic expression"
         where
             table =
                 [
@@ -51,35 +51,32 @@ instance Parse Aexp where
                 ]
             term =
                 choice
-                    [ parens (parse @Aexp) <?> "parenthesized arithmetic expression"
+                    [ parens (parses @Aexp) <?> "parenthesized arithmetic expression"
                     , Val <$> integer
                     , Var <$> identifier
-                    , Time <$ keyword "time" <*> parse @Stm
+                    , Time <$ keyword "time" <*> parses @Stm
                     ]
 
--- | Parser for boolean expressions, 'Bexp'.
--- Supports boolean literals, boolean operations, and relational expressions.
-instance Parse Bexp where
-    parse = buildExpressionParser table term <?> "boolean expression"
+-- | TODO
+instance Parses Bexp where
+    parses = buildExpressionParser table term <?> "boolean expression"
         where
             table =
                 [ [Prefix (Not <$ operator "not")]
                 , [Infix (And <$ operator "and") AssocLeft]
                 , [Infix (Or <$ operator "or") AssocLeft]
                 ]
-            relation = flip Rel <$> parse @Aexp <*> parse @Rop <*> parse @Aexp
             term =
                 choice
-                    [ parens (parse @Bexp) <?> "parenthesized boolean expression"
-                    , relation <?> "relation"
+                    [ parens (parses @Bexp) <?> "parenthesized boolean expression"
+                    , flip Rel <$> parses @Aexp <*> parses @Rop <*> parses @Aexp <?> "relation"
                     , Lit True <$ keyword "true"
                     , Lit False <$ keyword "false"
                     ]
 
--- | Parser for relational operators, 'Rop'.
--- Supports equality, inequality, and comparison operators.
-instance Parse Rop where
-    parse =
+-- | TODO
+instance Parses Rop where
+    parses =
         choice
             [ Eq <$ operator "="
             , Neq <$ operator "#"
@@ -89,126 +86,9 @@ instance Parse Rop where
             , Gt <$ operator ">"
             ]
 
--- | Parser for statements, 'Stm'. Supports all IMP language statements.
-instance Parse Stm where
-    parse = buildExpressionParser table term <?> "statement"
-        where
-            table =
-                [ [Infix (NonDet <$ keyword "[]") AssocLeft]
-                , [Infix (Par <$ keyword "par") AssocLeft]
-                , [Infix (Alternate <$ keyword "alternate") AssocLeft]
-                , [Infix (Seq <$ symbol ";") AssocLeft]
-                ]
-            term =
-                choice . map try $
-                    [ parens (parse @Stm) <?> "parenthesized statement"
-                    , Skip <$ keyword "skip"
-                    , VarDef <$> variable <*> parse @Dop <*> parse @Aexp
-                    , If
-                        <$ keyword "if"
-                        <*> parse @Bexp
-                        <* keyword "then"
-                        <*> parse @Stm
-                        <*> option Skip (keyword "else" *> parse @Stm)
-                        <* keyword "end"
-                    , While
-                        <$ keyword "while"
-                        <*> parse @Bexp
-                        <* keyword "do"
-                        <*> parse @Stm
-                        <* keyword "end"
-                    , Print <$ keyword "print" <*> parse @Aexp
-                    , Read
-                        <$ keyword "read"
-                        <*> identifier
-                    , Local
-                        <$ keyword "var"
-                        <*> identifier
-                        <* operator ":="
-                        <*> parse @Aexp
-                        <* keyword "in"
-                        <*> parse @Stm
-                        <* keyword "end"
-                    , fmap ProcDef $
-                        Procedure
-                            <$ keyword "procedure"
-                            <*> identifier
-                            <*> paramret
-                            <* keyword "begin"
-                            <*> parse @Stm
-                            <* keyword "end"
-                    , ProcInvoc
-                        <$> identifier
-                        <*> argret
-                    , (\s b -> Seq s $ While (Not b) s)
-                        <$ keyword "repeat"
-                        <*> parse @Stm
-                        <* keyword "until"
-                        <*> parse @Bexp
-                    , forto
-                        <$ keyword "for"
-                        <*> identifier
-                        <* operator ":="
-                        <*> parse @Aexp
-                        <* keyword "to"
-                        <*> parse @Aexp
-                        <* keyword "do"
-                        <*> parse @Stm
-                        <* keyword "end"
-                    , forto "_times" (Val 0) -- unassignable counter variable prevents modification from body
-                        <$ keyword "do"
-                        <*> parse @Aexp
-                        <* keyword "times"
-                        <*> parse @Stm
-                    , Revert
-                        <$ keyword "revert"
-                        <*> parse @Stm
-                        <* keyword "if"
-                        <*> parse @Bexp
-                    , Break <$ keyword "break" -- also parse outside while
-                    , Match
-                        <$ keyword "match"
-                        <*> parse @Aexp
-                        <* keyword "on"
-                        <*> many branch
-                        <* keyword "default"
-                        <* symbol ":"
-                        <*> parse @Stm
-                        <* keyword "end"
-                    , Havoc <$ keyword "havoc" <*> identifier
-                    , Assert <$ keyword "assert" <*> parse @Bexp
-                    , Flip
-                        <$ keyword "flip"
-                        <*> parens integer
-                        <*> parse @Stm
-                        <* keyword "flop"
-                        <*> parse @Stm
-                        <* keyword "end"
-                    , Raise <$ keyword "raise" <*> parse @Aexp
-                    , Try
-                        <$ keyword "try"
-                        <*> parse @Stm
-                        <* keyword "catch"
-                        <*> identifier
-                        <* keyword "with"
-                        <*> parse @Stm
-                        <* keyword "end"
-                    , Swap
-                        <$ keyword "swap"
-                        <*> identifier
-                        <*> identifier
-                    , Timeout
-                        <$ keyword "timeout"
-                        <*> parse @Stm
-                        <* keyword "after"
-                        <*> parse @Aexp
-                        <* keyword "end"
-                    ]
-
--- | Parser for variable definition operators, 'Dop'.
--- Supports simple, increment, decrement, product, quotient, and remainder definitions.
-instance Parse Dop where
-    parse =
+-- | TODO
+instance Parses Dop where
+    parses =
         choice
             [ Def <$ operator ":="
             , Inc <$ operator "+="
@@ -218,60 +98,161 @@ instance Parse Dop where
             , Rem <$ operator "%="
             ]
 
--- | Parser for IMP language constructs, 'Construct' which encapsulate
--- arithmetic and boolean expressions, statements and whitespace.
-instance Parse Construct where
-    parse =
+-- | TODO
+instance Parses Stm where
+    parses = buildExpressionParser table term <?> "statement"
+        where
+            table =
+                if extensions
+                    then
+                        [ [Infix (NonDet <$ operator "[]") AssocLeft]
+                        , [Infix (Par <$ operator "par") AssocLeft]
+                        , [Infix (Alternate <$ operator "alternate") AssocLeft]
+                        , [Infix (Seq <$ operator ";") AssocLeft]
+                        ]
+                    else [[Infix (Seq <$ operator ";") AssocLeft]]
+            term =
+                choice . map try $
+                    if extensions then imp ++ ext else imp
+
+imp :: [Parser Stm]
+imp =
+    [ parens (parses @Stm) <?> "parenthesized statement"
+    , Skip <$ keyword "skip"
+    , VarDef <$> variable <*> parses @Dop <*> parses @Aexp
+    , IfElse
+        <$ keyword "if"
+        <*> parses @Bexp
+        <* keyword "then"
+        <*> parses @Stm
+        <*> option Skip (keyword "else" *> parses @Stm)
+        <* keyword "end"
+    , While
+        <$ keyword "while"
+        <*> parses @Bexp
+        <* keyword "do"
+        <*> parses @Stm
+        <* keyword "end"
+    , Print <$ keyword "print" <*> parses @Aexp
+    , Read <$ keyword "read" <*> identifier
+    ]
+
+ext :: [Parser Stm]
+ext =
+    [ Local
+        <$ keyword "var"
+        <*> identifier
+        <* operator ":="
+        <*> parses @Aexp
+        <* keyword "in"
+        <*> parses @Stm
+        <* keyword "end"
+    , fmap ProcDef $
+        Procedure
+            <$ keyword "procedure"
+            <*> identifier
+            <*> parens (signature identifier identifier)
+            <* keyword "begin"
+            <*> parses @Stm
+            <* keyword "end"
+    , ProcInvoc
+        <$> identifier
+        <*> parens (signature (parses @Aexp) variable) -- allow placeholder in return variables
+    , (\s b -> Seq s $ While (Not b) s)
+        <$ keyword "repeat"
+        <*> parses @Stm
+        <* keyword "until"
+        <*> parses @Bexp
+    , for
+        <$ keyword "for"
+        <*> identifier
+        <* operator ":="
+        <*> parses @Aexp
+        <* keyword "to"
+        <*> parses @Aexp
+        <* keyword "do"
+        <*> parses @Stm
+        <* keyword "end"
+    , for "_times" (Val 0) -- unassignable counter variable prevents modification from body
+        <$ keyword "do"
+        <*> parses @Aexp
+        <* keyword "times"
+        <*> parses @Stm
+    , Revert
+        <$ keyword "revert"
+        <*> parses @Stm
+        <* keyword "if"
+        <*> parses @Bexp
+    , Break <$ keyword "break" -- INFO: parses outside while
+    , Match
+        <$ keyword "match"
+        <*> parses @Aexp
+        <* keyword "on"
+        <*> many branch
+        <* keyword "default"
+        <* symbol ":"
+        <*> parses @Stm
+        <* keyword "end"
+    , Havoc <$ keyword "havoc" <*> identifier
+    , Assert <$ keyword "assert" <*> parses @Bexp
+    , FlipFlop
+        <$ keyword "flip"
+        <*> parens integer
+        <*> parses @Stm
+        <* keyword "flop"
+        <*> parses @Stm
+        <* keyword "end"
+    , Raise <$ keyword "raise" <*> parses @Aexp
+    , TryCatch
+        <$ keyword "try"
+        <*> parses @Stm
+        <* keyword "catch"
+        <*> identifier
+        <* keyword "with"
+        <*> parses @Stm
+        <* keyword "end"
+    , Swap
+        <$ keyword "swap"
+        <*> identifier
+        <*> identifier
+    , Timeout
+        <$ keyword "timeout"
+        <*> parses @Stm
+        <* keyword "after"
+        <*> parses @Aexp
+        <* keyword "end"
+    ]
+
+-- | TODO
+instance Parses Construct where
+    parses =
         choice . map try $
-            [ Statement <$> parse @Stm
-            , Boolean <$> parse @Bexp
-            , Arithmetic <$> parse @Aexp
+            [ Statement <$> parses @Stm
+            , Boolean <$> parses @Bexp
+            , Arithmetic <$> parses @Aexp
             , Whitespace <$ whitespace
             ]
 
--- | Constructor for bounded loop.
-forto :: String -> Aexp -> Aexp -> Stm -> Stm
-forto x e1 e2 s =
-    Local x e1 $
-        -- stop condition is evaluated every iteration
+-- | TODO
+for :: String -> Aexp -> Aexp -> Stm -> Stm
+for x a1 a2 s =
+    Local x a1 $
         While
-            (Rel Lt (Var x) e2)
+            (Rel Lt (Var x) a2) -- stop condition is evaluated every iteration
             (Seq s $ VarDef x Inc (Val 1))
 
--- | Parser for procedure parameter and return variable lists.
-paramret :: Parser ([String], [String])
-paramret =
-    parens $
-        (,)
-            <$> sepBy identifier (symbol ",")
-            <* symbol ";"
-            <*> sepBy identifier (symbol ",")
+signature :: Parser a -> Parser b -> Parser ([a], [b])
+signature p1 p2 =
+    (,)
+        <$> sepBy p1 (symbol ",")
+        <* symbol ";"
+        <*> sepBy p2 (symbol ",")
 
--- | Parser for procedure argument and return variable lists.
-argret :: Parser ([Aexp], [String])
-argret =
-    parens $
-        (,)
-            <$> sepBy (parse @Aexp) (symbol ",")
-            <* symbol ";"
-            <*> sepBy variable (symbol ",") -- allow placeholders in returns
-
--- | Parser for single pattern match branch.
+-- | TODO
 branch :: Parser (Integer, Stm)
 branch =
     (,)
         <$> integer
         <* symbol ":"
-        <*> parse @Stm
+        <*> parses @Stm
         <* symbol ","
-
--- | Wrap string in angle brackets.
-wrap :: String -> String
-wrap = ('<' :) . (++ ">") -- this is why haskell is awesome
-
--- | Top-level parser for any type that implements 'Parse'.
--- Takes a channel name for error reporting and input string to parse,
--- and returns either parse error or the successfully parsed value.
--- Used in "IMP.REPL" to parse user input and "IMP.CLI" to parse source code.
-parser :: (Parse a) => String -> String -> Either ParseError a
-parser channel = Parsec.parse (whitespace *> parse <* eof) $ wrap channel
